requirements.txt
Flask==3.1.1
flask_cors==6.0.1
pandas==2.3.1
pymongo==4.6.3
python-dotenv==1.1.1
scikit_learn==1.4.1.post1
waitress==3.0.2
